safari-534.53-branch/LayoutTests/svg/dom/SVGAnimatedInteger-expected.txt

CONSOLE MESSAGE: line 13: TypeError: 'undefined' is not an object (evaluating 'filterElement.filterResX.baseVal')
This test checks the SVGAnimatedInteger API - utilizing the filterResX property of SVGFilterElement

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Check initial filterResX value
FAIL filterElement.filterResX.toString() should be [object SVGAnimatedInteger]. Threw exception TypeError: 'undefined' is not an object (evaluating 'filterElement.filterResX.toString')
FAIL typeof(filterElement.filterResX.baseVal) should be number. Threw exception TypeError: 'undefined' is not an object (evaluating 'filterElement.filterResX.baseVal')
FAIL filterElement.filterResX.baseVal should be 0. Threw exception TypeError: 'undefined' is not an object (evaluating 'filterElement.filterResX.baseVal')

Check that integers are static, caching value in a local variable and modifying it, should have no effect
FAIL successfullyParsed should be true. Threw exception ReferenceError: Can't find variable: successfullyParsed

TEST COMPLETE
